add condition check to prevent navcontroller from triggering previous navaction second time

+++ app/src/main/java/com/yaroslavm/cft/ui/response/BinResponse.kt
@@ -13,7 +13,7 @@ data class BinResponse(
 ) {
 
     data class Number(
-        @SerializedName("length") var length: Double? = null,
+        @SerializedName("length") var length: Int? = null,
         @SerializedName("luhn") var luhn: Boolean? = null,
     )
